Refactor multiscene blend functions to their own module
