+++ satpy/multiscene/__init__.py
@@ -1,2 +1,4 @@
+"""Functions and classes related to MultiScene functionality."""
+
+from ._blend_funcs import stack, timeseries  # noqa
 from ._multiscene import MultiScene  # noqa
-from ._multiscene import stack, timeseries  # noqa
